requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
email-validator==2.2.0
motor==3.6.0
beanie==1.27.0
pydantic==2.9.2
pydantic-settings==2.5.2
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.2
locust==2.31.5
python-dotenv==1.0.1
pymongo==4.9.1
PyJWT
passlib[bcrypt]
bcrypt==3.2.2